# hydranms/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HydraNMS multi-tenant network monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vpn/sites:
    get:
      operationId: getVpnSites
      tags: [vpn]
      summary: List WireGuard sites visible to the active company
      responses:
        "200":
          description: WireGuard site inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VpnSite"
    post:
      operationId: createVpnSite
      tags: [vpn]
      summary: Create a WireGuard site for a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VpnSiteInput"
      responses:
        "201":
          description: WireGuard site created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnSite"
        "400":
          description: Invalid or overlapping network
        "403":
          description: Super-admin access required
  /vpn/sites/{siteId}/bundle:
    post:
      operationId: generateVpnSiteBundle
      tags: [vpn]
      summary: Generate a one-time WireGuard and RouterOS onboarding bundle
      parameters:
        - name: siteId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: WireGuard configuration bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnSiteBundle"
        "404":
          description: Site not found in the active company
        "409":
          description: WireGuard server settings are not configured
  /vpn/sites/{siteId}/revoke:
    post:
      operationId: revokeVpnSite
      tags: [vpn]
      summary: Revoke a WireGuard site peer
      parameters:
        - name: siteId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Site peer revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnSite"
        "403":
          description: Super-admin access required
        "404":
          description: Site not found in the active company
...
components:
  schemas:
...
    VpnSite:
      type: object
      required: [id, companyId, name, lanCidr, tunnelAddress, routerOsVersion, status, routeState, lastHandshakeAt, lastBundleAt, revokedAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        name: { type: string }
        lanCidr: { type: string }
        tunnelAddress: { type: string }
        routerOsVersion: { type: string }
        status: { type: string, enum: [pending, active, offline, revoked] }
        routeState: { type: string, enum: [pending, applied, error] }
        lastHandshakeAt: { type: ["string", "null"], format: date-time }
        lastBundleAt: { type: ["string", "null"], format: date-time }
        revokedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    VpnSiteInput:
      type: object
      required: [companyId, name, lanCidr, routerOsVersion]
      properties:
        companyId: { type: string, minLength: 1 }
        name: { type: string, minLength: 1, maxLength: 120 }
        lanCidr: { type: string, pattern: "^[0-9.]+/[0-9]+$" }
        routerOsVersion: { type: string, minLength: 1, maxLength: 40 }
    VpnSiteBundle:
      type: object
      required: [siteId, name, tunnelAddress, lanCidr, serverEndpoint, serverPublicKey, clientPrivateKey, clientPublicKey, wireguardConfig, routerOsScript, serverPeerSnippet, generatedAt]
      properties:
        siteId: { type: string }
        name: { type: string }
        tunnelAddress: { type: string }
        lanCidr: { type: string }
        serverEndpoint: { type: string }
        serverPublicKey: { type: string }
        clientPrivateKey: { type: string }
        clientPublicKey: { type: string }
        wireguardConfig: { type: string }
        routerOsScript: { type: string }
        serverPeerSnippet: { type: string }
        generatedAt: { type: string, format: date-time }
...