# hydranms/lib/api-zod/src/generated/types/vpnSiteBundle.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydraNMS multi-tenant network monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface VpnSiteBundle {
  siteId: string;
  name: string;
  tunnelAddress: string;
  lanCidr: string;
  serverEndpoint: string;
  serverPublicKey: string;
  clientPrivateKey: string;
  clientPublicKey: string;
  wireguardConfig: string;
  routerOsScript: string;
  serverPeerSnippet: string;
  generatedAt: Date;
}