# hydranms/lib/api-zod/src/generated/types/vpnSiteInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydraNMS multi-tenant network monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface VpnSiteInput {
  /** @minLength 1 */
  companyId: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @pattern ^[0-9.]+/[0-9]+$ */
  lanCidr: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  routerOsVersion: string;
}